Fixed a sign error on the matrix.
For the future: the definition of the double layer has the opposite sign with respect to Helmut's.
The consequence was that the diagonal element of the second value of S_i_ was wrong. This has been fixed by changing
the sign for the factor in constructSi()

This implies that when porting the code from his module care has to be taken in order to avoid
sign problems. This bug was the same as a previous commit (6efa8aaedd...)

Extra printing functionality added in sparse2.