Working on unit tests and linking to math libs.
1. Keeping the unit tests up to speed. Various cosmetics in their source code and compilation.
2. We now avoid completely to link to math libs. They are not needed (as we use Eigen as math backend) in the
C/C++ code. PEDRA needed DSYEV LAPACK routine. This has been substituted with custom routines specialized
for 3x3 matrices.

IF the user wants to used the Eigen fallback mechanism to MKL, then we check for BLAS/LAPACK. Still
it is UP TO THE USER to check that the right versions of Eigen and MKL are being used and most of
all that CMake is finding MKL (and not Atlas for example)