Still working on the symmetric generation of the cavity.
There seems to be some problem when adding new spheres. The number of tesserae generated is different if working with symmetry.
Generated cavity looks fine when addition of spheres is turned off.
A further problem, that I don't quite understand, is that the plotcav subroutine writes a .off file containing only the
symmetry unique portion of the cavity, even though it is called -after- the replication has been done...